moves everything over to a new app called dataviz to get it out of the marketing app
