+++ app/app/settings.py
@@ -108,7 +108,11 @@ AUTHENTICATION_BACKENDS = (
 TEMPLATES = [
     {
         'BACKEND': 'django.template.backends.django.DjangoTemplates',
-        'DIRS': ['retail/templates/', 'external_bounties/templates/'],
+        'DIRS': [
+            'retail/templates/',
+            'external_bounties/templates/',
+            'dataviz/templates',
+        ],
         'APP_DIRS': True,
         'OPTIONS': {
             'context_processors': [
