+++ app/dataviz/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class DatavizConfig(AppConfig):
+    name = 'dataviz'
