+++ app/dataviz/templates/cohort.html
@@ -32,6 +32,7 @@
 {% endblock %}
 
 {% block content %}
+<a href="..">&lt; back</a>
 
 <style>
     td{
