+++ app/dataviz/templates/dataviz/graph.html
@@ -58,6 +58,7 @@
 
 
 <body>
+<a href="..">&lt; back</a>
 {% include "dataviz/shared/nav.html" %}
 
 <script>
