+++ app/dataviz/templates/dataviz/index.html
@@ -28,10 +28,11 @@
 
 {% block breadcrumbs %}{% endblock %}
 
-{% block title %}Stats | Gitcoin
+{% block title %}Analytics Playground | Gitcoin
 {% endblock %}
 
 {% block content %}
+<a href="..">&lt; back</a>
 <ul>
   <li>
     <a href='{% url 'viz_sunburst' %}'>Sunburst</a>
