+++ app/dataviz/templates/dataviz/shared/nav.html
