+++ app/dataviz/templates/dataviz/spiral.html
@@ -28,10 +28,11 @@
 
 {% block breadcrumbs %}{% endblock %}
 
-{% block title %}Stats | Gitcoin
+{% block title %}Spiral Graph | Gitcoin
 {% endblock %}
 
 {% block content %}
+<a href="..">&lt; back</a>
 <style>
       body {
         font-family: 'Lucida Grande', 'Lucida Sans Unicode', 'Geneva', 'Verdana', sans-serif;
@@ -85,6 +86,7 @@
       <link rel="stylesheet" type="text/css" href="style.css"/>
       <script data-require="d3@4.0.0" data-semver="4.0.0" src="https://d3js.org/d3.v4.js"></script>
   </head>
+  <a href="..">&lt; back</a>
   <body>
     <h2>{{key}}</h2>
     {% include "dataviz/shared/nav.html" %}
@@ -102,7 +104,7 @@
 
     {% endfor %}
     </script>
-    <script type="text/javascript" src="/static/v2/js/stats/spiral.js"></script>
+    <script type="text/javascript" src="/static/v2/js/dataviz/spiral.js"></script>
   </body>
 
 {% endblock %}
