+++ app/dataviz/templates/funnel.html
@@ -32,6 +32,7 @@
 {% endblock %}
 
 {% block content %}
+<a href="..">&lt; back</a>
 
 <style>
 td.raw{ border-left: 1px #ccc solid; }
