+++ app/retail/templates/admin/index.html
@@ -71,9 +71,9 @@
 
 <div id="content-related">
     <div class="module" id="-some-custom-links">
-        <h2>{% trans 'Stats' %}</h2>
+        <h2>{% trans 'Analytics' %}</h2>
         <ul>
-            <a href="{% url 'stats' %}">Stats</a>
+            <a href="{% url 'stats' %}">KPI Dashboard</a>
         </ul>
         <ul>
             <a href="{% url 'cohort' %}">Cohort Analysis</a>
@@ -82,7 +82,7 @@
             <a href="{% url 'funnel' %}">Funnel Analysis</a>
         </ul>
         <ul>
-            <a href="/_administration/viz/">D3 Data Visualizations</a>
+            <a href="/_administration/viz/">Analytics Playground</a>
         </ul>
         <h2>{% trans 'Administrative Tasks' %}</h2>
